python-backend/requirements.txt
boto3
mysql-connector-python
scikit-learn
joblib
